refactor/#85: authentication 인증 절차 filter chain으로 자동화
