+++ src/main/java/com/epi/epilog/app/controller/DiabetesController.java
@@ -3,8 +3,8 @@ package com.epi.epilog.app.controller;
 import com.epi.epilog.app.dto.CommonResponseDto;
 import com.epi.epilog.app.dto.DiabetesRequestDto;
 import com.epi.epilog.app.dto.DiabetesResponseDto;
-import com.epi.epilog.app.service.diabetes.DiabetesCommandService;
-import com.epi.epilog.app.service.diabetes.DiabetesQueryService;
+import com.epi.epilog.app.service.logs.DiabetesCommandService;
+import com.epi.epilog.app.service.logs.DiabetesQueryService;
 import com.epi.epilog.global.exception.ApiException;
 import com.epi.epilog.global.exception.ErrorCode;
 import com.epi.epilog.global.utils.CustomUserDetails;
