+++ src/main/java/com/epi/epilog/app/controller/LogController.java
@@ -1,7 +1,7 @@
 package com.epi.epilog.app.controller;
 
 import com.epi.epilog.app.dto.LogsResponseDto;
-import com.epi.epilog.app.service.LogQueryService;
+import com.epi.epilog.app.service.logs.LogQueryService;
 import com.epi.epilog.global.exception.ApiException;
 import com.epi.epilog.global.exception.ErrorCode;
 import com.epi.epilog.global.utils.CustomUserDetails;
@@ -29,11 +29,7 @@ public class LogController {
     public LogsResponseDto.MonthLogsCount monthLogsCount(@RequestParam(value = "date", required = false) String date){
         try {
             LocalDate queryDate;
-
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)) {
-                throw new ApiException(ErrorCode.INVALID_TOKEN);
-            }
 
             if (date == null) {
                  date  = DateTimeConverter.convertLocalDateToString(LocalDate.now());
@@ -55,9 +51,6 @@ public class LogController {
     public LogsResponseDto.DayLogsList dayLogList(@RequestParam(value = "date", required = false) String date){
         try{
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)){
-                throw new ApiException(ErrorCode.INVALID_TOKEN);
-            }
 
             if (date == null){
                 date = DateTimeConverter.convertLocalDateToString(LocalDate.now());
@@ -79,9 +72,6 @@ public class LogController {
     public LogsResponseDto.DayAvgBloodSugar dayAvgBloodSugar(@RequestParam(value = "date", required = false)String date){
         try {
             Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)){
-                throw new ApiException(ErrorCode.INVALID_TOKEN);
-            }
 
             if (date == null){
                 date = DateTimeConverter.convertLocalDateToString(LocalDate.now());
@@ -101,13 +91,10 @@ public class LogController {
     @GetMapping("/bloodsugar")
     public LogsResponseDto.DayBloodSugarList getDayBloodSugarList(@RequestParam(value = "date", required = false)String date){
         try {
-            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-            if (authentication == null || !(authentication.getPrincipal() instanceof CustomUserDetails)){
-                throw new ApiException(ErrorCode.INVALID_TOKEN);
-            }
             if (date == null) {
                 date = DateTimeConverter.convertLocalDateToString(LocalDate.now());
             }
+            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
             return logQueryService.dayBloodSugarList(((CustomUserDetails) authentication.getPrincipal()).getMember(), date);
         }catch (Exception e){
             throw new ApiException(ErrorCode.INTERNAL_SERVER_ERROR, e);
@@ -117,6 +104,16 @@ public class LogController {
     /**
      * 월별 체중, 체지방률 목록 조회
      */
+//    @GetMapping("/weight")
+//    public LogsResponseDto.MonthWeightList getMonthWeightList(@RequestParam(value="date", required = false)String date){
+//        try {
+//            Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+//
+//
+//        } catch (Exception e) {
+//            throw new ApiException(ErrorCode.INTERNAL_SERVER_ERROR, e);
+//        }
+//    }
 
     /**
      * 일지 등록
