+++ src/main/java/com/epi/epilog/global/config/SecurityConfig.java
@@ -47,7 +47,8 @@ public class SecurityConfig {
 
         // usernamePasswordAuthenticationToken 앞에 jwt 필터체인 추가
         http.addFilterBefore(new JwtAuthFilter(jwtUtil, customUserDetailService), UsernamePasswordAuthenticationFilter.class);
-
+        // AuthenticationFilterChain 추가
+        http.addFilterBefore(new CustomAuthenticationFilter(), UsernamePasswordAuthenticationFilter.class);
         // 권한 규칙 설정
         http.authorizeHttpRequests(authorize -> authorize
                 .requestMatchers(AUTH_WHITELIST).permitAll()
